Fixed printf arguments when regoff_t is not of type int.
When configured with --enable-system-abi the `rm_so' and `rm_eo'
fields of `struct regmatch_t' may not be of type `int'.  This was
actually causing agrep and the test programs to crash on some
platforms.

darcs-hash:20060521070435-ced27-eb905d44d2df0d0ce7188bd831bdc4b25a0231ea.gz

+++ src/agrep.c
@@ -414,12 +414,12 @@ tre_agrep_handle_file(const char *filename)
 
 	      if (color_option && !invert_match)
 		{
-		  printf("%.*s", pmatch[0].rm_so, record);
+		  printf("%.*s", (int)pmatch[0].rm_so, record);
 		  printf("\33[%sm", highlight);
-		  printf("%.*s", pmatch[0].rm_eo - pmatch[0].rm_so,
+		  printf("%.*s", (int)(pmatch[0].rm_eo - pmatch[0].rm_so),
 			 record + pmatch[0].rm_so);
 		  fputs("\33[00m", stdout);
-		  printf("%.*s", record_len - pmatch[0].rm_eo,
+		  printf("%.*s", (int)(record_len - pmatch[0].rm_eo),
 			 record + pmatch[0].rm_eo);
 		}
 	      else
